+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Duplicated structured data in product search context
+
 ## [3.133.0] - 2024-08-01
 
 ### Added